server/requirements.txt
flask
pillow
numpy
requests
gunicorn
python-dotenv